move util function to core utils
